refractor: add types
