+++ src/components/CharacterCard/CharacterCard.tsx
@@ -1,19 +1,8 @@
 import React from 'react';
 import './CharacterCard.css';
+import { Character } from '../../types/Character';
 
-interface Character {
-  id: number;
-  name: string;
-  description: string;
-  imageUrl: string;
-  link: { [id: number]: string };
-}
-
-interface CharacterCardProps {
-  character: Character;
-}
-
-const CharacterCard: React.FC<CharacterCardProps> = ({ character }) => {
+const CharacterCard: React.FC<{ character: Character }> = ({ character }) => {
   return (
     <div className="character-card">
       <img src={character.imageUrl} alt={character.name} />
