+++ src/types/Character.ts
@@ -1,0 +1,7 @@
+export interface Character {
+  id: number;
+  name: string;
+  description: string;
+  imageUrl: string;
+  link: { [id: number]: string };
+}
