+++ src/types/SearchProps.ts
@@ -1,0 +1,4 @@
+export interface SearchProps {
+  onSearch: (query: string) => void;
+  count: string;
+}
